# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api TODO Documentation",
        "description": "Todo APP",
        "contact": {
            "email": "mohamed.amine.aloui@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/user/todos": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get List Of Todos",
                "description": "Returns all todos created by currently authenticated user",
                "operationId": "getAllTodos",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "addTodo"
                ],
                "summary": "Adding TODO",
                "description": "Adding a new Todo",
                "operationId": "addingNewTodo",
                "responses": {
                    "201": {
                        "description": "Todo Added Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "200": {
                        "description": "Validation Errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/user/todos/{id}": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get Todo by ID",
                "description": "Returns Todo by its id",
                "operationId": "getTodoByID",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todo"
                ],
                "summary": "Update Todo",
                "description": "Todo Updating",
                "operationId": "updateTodo",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "200": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todo"
                ],
                "summary": "Delete Todo",
                "description": "Delete Todo By ID",
                "operationId": "deleteTodo",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "userRegistration",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "200": {
                        "description": "Validation Errors : email already taken or field error or missed"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Logged User Data Using Auth Token",
                "description": "Returns Logged in user data using Auth Token",
                "operationId": "getLoggedInUserData",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "User login by email and password",
                "operationId": "userLogin",
                "responses": {
                    "200": {
                        "description": "LoggedIn",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found or Login Failed : Invalid Credentials"
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "logout"
                ],
                "summary": "User Logout",
                "description": "Logout User and revoke assigned Token",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Successful operation : User Logged out",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "User already logged-out"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}